Check channel address when receiving payment in CLI
When an IOU is received its signature covers a channel address. There
are two possible ways to proceed:

1. Take this address and apply the IOU to the channel with this addres.
2. Ask the receiving user to provide the address of the channel that
   this IOU has been received from.

Had we gone with option 1, we would not have this issue.

But we chose option 2, ironically, as an extra precaution: by explicitly
giving the address of the channel “through” which the IOU was received,
the user explicitly confirms that they know what they are doing. In this
case we must check that the address that the user has given (and the we
use to locate the state of the channel) matches the one in the IOU.